feature: utils.SnapshotQueue - added a new queue class
Since we cannot peek at the queue, we need to find some way to get the info inside of the queue. Since queue is actually a subclass of deque, which has deque.queue() method which returns a live copy of the queue, we can cast the return value of the method to list, effectively creating a snapshot of the queue.

To ensure thread safety, we use mutex that is already present inside of the queue.

https://stackoverflow.com/a/65098964